requirements.txt
----------------------------- -------
asgiref                       3.8.1
Django                        5.1.2
djangorestframework           3.15.2
djangorestframework-simplejwt 5.3.1
drf-yasg                      1.21.8
inflection                    0.5.1
packaging                     24.1
pip                           24.2
PyJWT                         2.9.0
pytz                          2024.2
PyYAML                        6.0.2
sqlparse                      0.5.1
tzdata                        2024.2
uritemplate                   4.1.1 